First watch face to feature the SUSE Chameleon
This is a watch face for the Fitbit Versa, Versa 2, or Versa Lite
smart watch.
The archive is exported directly from fitbit studio and contains
all the relevant files.

Note that the Chameleon logo is a registered trademark of SUSE LLC
and can be found at https://brand.suse.com/